fixing errors

+++ src/Servo_Functions/Tail_Function.cpp
@@ -6,7 +6,7 @@ static ServoBus* SB=nullptr; static Tail::Map MAP;
 namespace Tail {
 void begin(ServoBus* bus, const Map& map) {
   SB=bus; MAP=map;
-  SB->attach(MAP.wag, ServoLimits{500,2500,0,180});
+SB->attach(MAP.wag, ServoLimits{ (uint16_t)500, (uint16_t)2500, 0.0f, 180.0f });
 }
 void set(float amt01) {
   SB->writeDegrees(MAP.wag, 90 + 30*constrain(amt01,0,1));
